fixed event_id intake

+++ app/models/event_image.py
@@ -12,7 +12,7 @@ class EventImage(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     url = db.Column(db.String(255), nullable=False)
-    event_id = db.Column(db.Integer(2500), db.ForeignKey(add_prefix_for_prod("events.id")))
+    event_id = db.Column(db.Integer(), db.ForeignKey(add_prefix_for_prod("events.id")))
     owner_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod("users.id")))
